Removed the API version package and moved it to the root of endpoint
