+++ src/main/java/dev/orion/api/endpoint/dto/AddUserToActivityRequestDtoV1.java
@@ -1,0 +1,8 @@
+package dev.orion.api.endpoint.dto;
+
+import javax.validation.constraints.NotBlank;
+
+public class AddUserToActivityRequestDtoV1 {
+    @NotBlank
+    public String userExternalId;
+}
