+++ src/main/java/dev/orion/api/endpoint/dto/CreateActivityRequestDtoV1.java
@@ -1,0 +1,16 @@
+package dev.orion.api.endpoint.dto;
+
+import javax.validation.constraints.NotBlank;
+
+public class CreateActivityRequestDtoV1 {
+    @NotBlank(message = "User may not be blank")
+    private String userExternalId;
+
+    public String getUserExternalId() {
+        return userExternalId;
+    }
+
+    public void setUserExternalId(String userExternalId) {
+        this.userExternalId = userExternalId;
+    }
+}
