--- src/main/java/dev/orion/api/endpoint/v1/dto/AddUserToActivityRequestDtoV1.java
@@ -1,8 +1,0 @@
-package dev.orion.api.endpoint.v1.dto;
-
-import javax.validation.constraints.NotBlank;
-
-public class AddUserToActivityRequestDtoV1 {
-    @NotBlank
-    public String userExternalId;
-}
