+++ src/main/java/dev/orion/data/entity/Activity.java
@@ -23,7 +23,7 @@ public class Activity extends PanacheEntityBase {
     @JsonIgnore
     public Document document;
 
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = "activity")
+    @OneToMany(mappedBy = "activity", cascade = CascadeType.ALL)
     @OrderColumn
     @JsonManagedReference
     public Set<User> userList = new LinkedHashSet<>();
@@ -41,6 +41,10 @@ public class Activity extends PanacheEntityBase {
     LocalDateTime createdAt;
 
     LocalDateTime updatedAt;
+
+    public Set<User> getUserList() {
+        return userList;
+    }
 
     @PrePersist
     void createdAtUpdate() {
